feat(links): tidy code, show attachment errors on frontend
